+++ astropy/coordinates/angles.py
@@ -17,7 +17,7 @@ from .. import units as u
 from ..utils import deprecated
 
 
-__all__ = ['Angle', 'AngularSeparation', 'Latitude', 'Longitude']
+__all__ = ['Angle', 'Latitude', 'Longitude']
 
 
 TWOPI = math.pi * 2.0  # no need to calculate this all the time
@@ -604,79 +604,6 @@ class Longitude(Angle):
     def wrap_angle(self, value):
         self._wrap_angle = Angle(value)
         self._wrap_internal()
-
-
-class AngularSeparation(Angle):
-    """
-    An on-sky separation between two directions.
-
-    .. note::
-        This is computed using the Vincenty great circle distance
-        formula, and hence should be numerically stable even for
-        near antipodal points.
-
-    Parameters
-    ----------
-    lon1 : float
-        The value of the first longitudinal/azimuthal angle.
-
-    lat1 : float
-        The value of the first latitudinal/elevation angle.
-
-    lon2 : float
-        The value of the second longitudinal/azimuthal angle.
-
-    lat2 : float
-        The value of the second latitudinal/elevation angle.
-
-    unit : `~astropy.units`
-        The unit of the given angles.
-    """
-    def __new__(cls, lon1, lat1, lon2, lat2, unit,
-        _supresslatlonswap_warning=False):  # TODO: remove this parameter in v0.4
-        # TODO: remove this warning in v0.4
-        if not _supresslatlonswap_warning:
-            from warnings import warn
-            from ..utils.exceptions import AstropyBackwardsIncompatibleChangeWarning
-            warn(AstropyBackwardsIncompatibleChangeWarning('The ordering of '
-                ' the AngularSeparation initializer angles was changed '
-                'from lat1, lon1, lat2, lon2 in v0.2 to "lon1, lat1, lon2, '
-                'lat2" in v0.3.  You MUST update your code to swap lat/lon '
-                'if you are not using keywords, or you will get the wrong '
-                'result.'))
-
-        unit = cls._convert_unit_to_angle_unit(unit)
-        lon1 = unit.to(u.radian, lon1)
-        if lat1 or lat2 or lon2:
-            lat1 = unit.to(u.radian, lat1)
-            lon2 = unit.to(u.radian, lon2)
-            lat2 = unit.to(u.radian, lat2)
-
-            sepval = util.vincenty_sphere_dist(lon1, lat1, lon2, lat2)
-        else:  # this is the case where lat1, lat2, and lon2 are all 0 or None or False
-            sepval = lon1
-
-        self = super(AngularSeparation, cls).__new__(cls, sepval, u.radian)
-
-        return self
-
-    def __quantity__(self):
-        unit = self._convert_unit_to_angle_unit(unit)
-        if unit is not None and unit.is_equivalent(u.radian):
-            return AngularSeparation
-        return Quantity
-
-    def __add__(self, other):
-        raise TypeError('+ is ambiguous for AngularSeparation objects; not supported')
-
-    def __radd__(self, other):
-        raise TypeError('+ is ambiguous for AngularSeparation objects; not supported')
-
-    def __sub__(self, other):
-        raise TypeError('- is ambiguous for AngularSeparation objects; not supported')
-
-    def __rsub__(self, other):
-        raise TypeError('- is ambiguous for AngularSeparation objects; not supported')
 
 
 #<----------------------------------Rotations---------------------------------->